% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFreqs_mae.R
\name{calcFreqs_mae}
\alias{calcFreqs_mae}
\title{calcFreqs_mae}
\usage{
calcFreqs_mae(mae, x1, x2, y1 = NULL, y2 = NULL, percY1 = 1/3, percY2 = 2/3)
}
\arguments{
\item{mae}{MultiAssayExperiment object containing methylation and expression matrices.}

\item{x1, x2}{Coordinates of vertical points in the X axis. Because it is expected to contain methylation values that vary between 0 and 1, the default values are 1/3 and 2/3.}

\item{y1, y2}{Coordinates of vertical points in the Y axis. Leaving them as NULL assigns them the percentiles of yVec defined by `percY1` and `percY2`.}

\item{percY1, percY2}{Values used to act as default for `y1` and `y2` when these are set to `NULL`}
}
\value{
a matrix with calculated frequencies
}
\description{
\code{calcFreqs_mae} Given a MultiAssayExperiment object with methylation and expression matrices, this function overimposes a grid on the scatterplot defined by YMet~Xmet
#' and returns a 3X3 matrix of counts with the number of points in each cell of the grid
#' for given vertical and horizontal lines.
}
\examples{

mae <- MultiAssayExperiment::MultiAssayExperiment(
  experiments = list(methylation = matrix(runif(1000), nrow=100), 
  expression = matrix(rnorm(1000), nrow=100))
)
x1 <- 1/3
x2 <- 2/3
y1 <- NULL
y2 <- NULL
percY1 <- 1/3
percY2 <- 2/3
calcFreqs_mae(mae, x1, x2, y1, y2, percY1, percY2)

}
\keyword{calculation}
\keyword{frequencies}
