% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationSelection.R
\name{correlationSelection}
\alias{correlationSelection}
\title{correlationSelection: A vector correlation function calculator}
\usage{
correlationSelection(
  X,
  Y,
  type = "Spearman",
  adj = TRUE,
  pValCutoff = 0.05,
  rCutoff = 0,
  sortByCorrs = FALSE
)
}
\arguments{
\item{X}{First matrix}

\item{Y}{Second matrix. Must have the same dimensions as X.}

\item{type}{specifies the correlation to choose between Spearman and Pearson. Default is Spearman.}

\item{adj}{logical variable indicating if the p-value returned should be adjusted or not. Default set to TRUE, which will return an adjusted p-value.}

\item{pValCutoff}{the upper limit to be used for the  p-value. Default is 0.05.}

\item{rCutoff}{the upper limit to be used for the correlation coefficient. Default is 0, no cut off.}

\item{sortByCorrs}{logical; if TRUE, results will be ordered in ascending order by p-value. Default set to FALSE.}
}
\value{
dataframe with correlations selected
}
\description{
\code{correlationSelection} Uses the function matCorrs; given two matrices X (m,n) , Y (m,n) this function computes Pearson and Spearman correlation coefficients
and their significance p-values  for every pair of row vectors.
}
\examples{

(X <- round(matrix (rnorm(30)*10, ncol=6),1))
(Y <- round(X + matrix (rnorm(30)*10, ncol=6),1))
(rownames(X)=rownames(Y)=letters[1:nrow(X)])
correlationSelection(X, Y, pValCutoff=0.25, rCutoff=0.1, type="Spearman", sortByCorrs=TRUE)


}
\keyword{Calculator}
\keyword{Correlation}
\keyword{Selection}
