% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenesLocations.R
\name{getGenesLocations}
\alias{getGenesLocations}
\title{A transcript coordinate annotation tool}
\usage{
getGenesLocations(geneSymbolsSEL, sortByChrom = TRUE, csvFileName = NULL)
}
\arguments{
\item{geneSymbolsSEL}{vector containing Gene Symbols (unique abbreviation for the gene name) to be annotated}

\item{sortByChrom}{logical; indicating if the results have to be ssorted ascending by chromosome number. Default to TRUE}

\item{csvFileName}{default is NULL; if a name is given, a csv file will be written as output}
}
\description{
\code{getGenesLocations} Given a vector with Gene Symbols, it will produce an object with ENTREZ ID, gene name, chromosome number,
start position and end position for Homo sapiens.
}
\examples{
\dontrun{
transcriptCoords <- getGenesLocations(rownames(selectedGenes),
csvFileName="results/coordsselectedGenes.csv")
}

}
\keyword{annotation}
\keyword{gene}
