% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mae_calcFreqs.R
\name{mae_calcFreqs}
\alias{mae_calcFreqs}
\title{mae_calcFreqs}
\usage{
mae_calcFreqs(
  mae,
  geneNum,
  x1,
  x2,
  y1 = NULL,
  y2 = NULL,
  percY1 = 1/3,
  percY2 = 2/3
)
}
\arguments{
\item{mae}{MultiAssayExperiment object containing methylation and expression matrices.}

\item{geneNum}{row of expression/methylation matrix for which the frequencies will be computed}

\item{x1, x2}{Coordinates of vertical points in the X axis. Because it is expected to contain methylation values that vary between 0 and 1, the default values are 1/3 and 2/3.}

\item{y1, y2}{Coordinates of vertical points in the Y axis. Leaving them as NULL assigns them the percentiles of yVec defined by `percY1` and `percY2`.}

\item{percY1, percY2}{Values used to act as default for `y1` and `y2` when these are set to `NULL`}
}
\value{
a matrix with calculated frequencies
}
\description{
\code{mae_calcFreqs} Given a MultiAssayExperiment object with methylation and expression matrices, 
this function overimposes a grid on the scatterplot defined by YMet~Xmet layers
and returns a 3X3 matrix of counts with the number of points in each cell of the grid
for given vertical and horizontal lines.
}
\examples{

# Methylation data 
methylData = matrix(runif(50), nrow=10)
colnames(methylData) <- paste0("samp", 1:ncol(methylData))
rownames(methylData) <- paste0("gene", 1:nrow(methylData))
# Expression data
expresData = matrix(rnorm(50), nrow=10)
colnames(expresData) <- paste0("samp", 1:ncol(methylData))
rownames(expresData) <- paste0("gene", 1:nrow(methylData))
# ColData
colDat <- data.frame(sampleID=colnames(methylData), 
                     name=letters[1:ncol(methylData)])
                     
rownames(colDat)<- colDat$sampleID
mae <- MultiAssayExperiment::MultiAssayExperiment(
experiments = list(methylation = methylData,
  expression = expresData),
  colData =colDat
)
geneRow <- 1
x1 <- 1/3 
x2 <- 2/3
y1 <- NULL
y2 <- NULL
percY1 <- 1/3
percY2 <- 2/3

mae_calcFreqs(mae=mae, geneNum=geneRow,
              x1, x2, y1, y2, percY1, percY2)
}
\keyword{calculation}
\keyword{frequencies}
