% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matCorrs.R
\name{matCorrs}
\alias{matCorrs}
\title{A row-wise correlation function calculator}
\usage{
matCorrs(X, Y)
}
\arguments{
\item{X}{First matrix}

\item{Y}{Second matrix. Must have the same dimensions as X}
}
\value{
matrix with a correlation value for each gene
}
\description{
\code{matCorrs} Given two matrices X (m,n), Y(m,n) this function computes (m) Pearson and Spearman correlation coefficients
and their significance p-values for every pair of row vectors.
}
\examples{
(X <- round(matrix (rnorm(30)*10, ncol=6),1)) + 1:10
(Y <- round(X + matrix (rnorm(30)*10, ncol=6),1)) - 10:1
(rownames(X)=rownames(Y)=letters[1:nrow(X)])
(m1<-matCorrs(X,Y))
matCorrs


}
\keyword{correlation}
