% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messageTitle.R
\name{messageTitle}
\alias{messageTitle}
\title{messageTitle
\code{messageTitle} A  wrapper function for sending messages to console.}
\usage{
messageTitle(aMessage, underChar = "-")
}
\arguments{
\item{aMessage}{text of the message.}

\item{underChar}{character to use for underlining the message.}
}
\description{
messageTitle
\code{messageTitle} A  wrapper function for sending messages to console.
}
\examples{
\dontrun{
messageTitle("Hello world", "$")
}

}
