% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneByName.R
\name{plotGeneByName}
\alias{plotGeneByName}
\title{plotGeneByName}
\usage{
plotGeneByName(
  geneName,
  mets,
  expmat,
  filename = NULL,
  text4Title = NULL,
  plotGrid = TRUE,
  figs = c(2, 2)
)
}
\arguments{
\item{geneName}{The name of the gene to be plotted.}

\item{mets}{A matrix containing the methylation data for the specified gene.}

\item{expmat}{A matrix containing the expression data for the specified gene.}

\item{filename}{If provided, the name of the file to save the results
as a PDF;  defaults to NULL.}

\item{text4Title}{A string used as the main title for the plot. Defaults to 
`geneName` if set to NULL.}

\item{plotGrid}{A boolean parameter indicating whether to pass
the grid option to the `plotGeneSel` function.}

\item{figs}{A two-component vector defining the 2-dimensional structure of
the plots to be generated.}
}
\value{
a pdf with scatterplots of selected by gene name
}
\description{
\code{plotGeneByName} plots points on a scatterplot with a 3x3 grid
superimposed. The name of a the gene is provided jointly with the matrix
and used to select the row to be plotted.
}
\examples{
# Plot gene by name based on example data
mets <- matrix(runif(1000), nrow = 100)
expres <- matrix(rnorm(1000), nrow = 100)
rownames(mets) <- paste0("Gene", 1:nrow(mets))
rownames(expres) <- paste0("Gene", 1:nrow(expres))
plotGeneByName(gene = "Gene40", mets = mets, expmat = expres)

}
\keyword{gene}
\keyword{name}
\keyword{plot}
