% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSel.R
\name{plotGeneSel}
\alias{plotGeneSel}
\title{plotGeneSel}
\usage{
plotGeneSel(
  xMet,
  yExp,
  titleText,
  x1 = 1/3,
  x2 = 2/3,
  y1 = NULL,
  y2 = NULL,
  percY1 = 1/3,
  percY2 = 2/3,
  plotGrid = TRUE
)
}
\arguments{
\item{xMet}{vector with methylation data.}

\item{yExp}{vector for expression data.}

\item{titleText}{plot title.}

\item{x1, x2}{Coordinates of vertical points in the X axis. Because it is expected to contain methylation values that vary between 0 and 1 the default values are 1/3 and 2/3.}

\item{y1, y2}{Coordinates of vertical points in the Y axis. Leaving them as NULL assigns them the percentiles of yVec defined by `percY1` and `percY2`.}

\item{percY1, percY2}{Values used to act as default for `y1`and `y2` when these are set to `NULL`}

\item{plotGrid}{logical. Defautl to TRUE will plot gridlines over the scatterplot.}
}
\description{
\code{plotGeneSel} plots points on a scatterplot with a 3x3 grid overimposed.
}
\examples{
xMet <- rnorm(100)
yExp <- rnorm(100)
titleText <- "Methylation-Gene Expression Correlation"
plotGeneSel(xMet, yExp, titleText)

}
\keyword{gene}
\keyword{plot}
\keyword{selection}
