% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenesMat.R
\name{plotGenesMat}
\alias{plotGenesMat}
\title{plotGenesMat}
\usage{
plotGenesMat(
  mets,
  expres,
  fileName = NULL,
  text4Title = NULL,
  x1 = 1/3,
  x2 = 2/3,
  y1 = NULL,
  y2 = NULL,
  percY1 = 1/3,
  percY2 = 2/3,
  plotGrid = TRUE,
  logicSc = NULL
)
}
\arguments{
\item{mets}{A matrix containing methylation data.}

\item{expres}{A matrix containing expression data.}

\item{fileName}{The name of the file used to save the results as a PDF.
If NULL, the plot is displayed on the screen.}

\item{text4Title}{An optional title for the plot, incorporating the gene name
and L-shape score. Defaults to NULL.}

\item{x1}{The x-coordinate of vertical points on the X-axis. 
Expected to contain methylation values ranging between 0 and 1,
with default values set to 1/3 and 2/3.}

\item{x2}{The x-coordinate of vertical points on the X-axis.
Expected to contain methylation values ranging between 0 and 1,
with default values set to 1/3 and 2/3.}

\item{y1}{The y-coordinate of vertical points on the Y-axis.
If NULL, these are set to the percentiles of `yVec` 
defined by `percY1` and `percY2`.}

\item{y2}{The y-coordinate of vertical points on the Y-axis.
If NULL, these are set to the percentiles of `yVec` 
defined by `percY1` and `percY2`.}

\item{percY1}{Values used as defaults for `y1` when it is set to NULL.}

\item{percY2}{Values used as defaults for `y2` when it is set to NULL.}

\item{plotGrid}{A logical value; defaults to TRUE, indicating 
whether to plot gridlines on the graph.}

\item{logicSc}{A numeric score representing the L-shape score. 
Defaults to NULL.}
}
\value{
a pdf with scatterplots for all genes
}
\description{
\code{plotGenesMat} wrapper function for plotting the scatterplots associated
with two matrices.
}
\examples{
mets <- matrix(runif(1000), nrow = 100)
expres <- matrix(rnorm(1000), nrow = 100)
rownames(mets) <- paste0("Gene", 1:nrow(mets))
rownames(expres) <- paste0("Gene", 1:nrow(expres))
# plotGenesMat (mets=mets, expres=expres, fileName = 'PlotAllGenes.pdf')

}
\keyword{gene}
\keyword{matrix}
\keyword{plot}
\keyword{scatterplot}
