% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read2.R
\name{read2}
\alias{read2}
\title{read2
\code{read2} Wrapper to read 2 files with the same format, dimensions and row and column names.}
\usage{
read2(expresFName, metFName, dataDirectory = ".", sepChar = ";",
  decChar = ".")
}
\arguments{
\item{expresFName}{Name of first file, expected to contain expression values.}

\item{metFName}{Name of second file, expected to contain methylation values.}

\item{dataDirectory}{Name of directory where the files are staore. Defaults to ".".}

\item{sepChar}{Name of character used to separate matrix columns. Defaults to ";".}

\item{decChar}{Name of character used as decimal point. Defaults to ".".}
}
\description{
read2
\code{read2} Wrapper to read 2 files with the same format, dimensions and row and column names.
}
\examples{
\dontrun{
(myData <- read2(myExpr, myMet))
}
}
