% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGAexpression}
\alias{TCGAexpression}
\title{Expression data matrix obtained from TCGA, TCGA-COAD dataset}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 30 columns.
}
\source{
\url{TCGA.org}
}
\usage{
data(TCGAexpression)
}
\value{
data matrix with gene expression data
}
\description{
A mxn matrix containing normalized gene expressions from TCGA
}
\keyword{datasets}
