% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGAmet}
\alias{TCGAmet}
\title{Methylation data matrix obtained from TCGA, TCGA-COAD dataset}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 30 columns.
}
\source{
\url{TCGA.org}
}
\usage{
data(TCGAmet)
}
\description{
A mxn matrix containing normalized methylation values from TCGA
}
\keyword{datasets}
