% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toReqMat.R
\name{toReqMat}
\alias{toReqMat}
\title{toReqMat
\code{toReqMat} can be used to turn a matrix of required percentages into
a matrix of required counts to facilitate its scoring}
\usage{
toReqMat(numPoints, aReqPercentMat)
}
\arguments{
\item{numPoints}{Number of points in a scatterplot. Used to turn the required percentages into required counts.}

\item{aReqPercentMat}{Matrix of required percentages}
}
\description{
toReqMat
\code{toReqMat} can be used to turn a matrix of required percentages into
a matrix of required counts to facilitate its scoring
}
\examples{
\dontrun{

}

}
