% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneByName.R
\name{plotGeneByName}
\alias{plotGeneByName}
\title{plotGeneByName}
\usage{
plotGeneByName(
  geneName,
  mets,
  expresmatrix,
  filename,
  text4Title = NULL,
  plotGrid = TRUE,
  figs = c(2, 2)
)
}
\arguments{
\item{geneName}{name of the gene to be plotted.}

\item{mets}{matrix containing the methylation data of the named gene.}

\item{expresmatrix}{containing the expression data of the named gene.}

\item{filename}{NULL, name of the file to store the results as pdf if a name is passed.}

\item{text4Title}{String to be used as main title for the plot. If NULL it is set to geneName}

\item{plotGrid}{boolean parameter to be passed to `plotGeneSel` function}

\item{figs}{2-components vector defining the 2-dimensional structure of the plots that will be drawn}
}
\description{
\code{plotGeneByName} plots points on a scatterplot with a 3x3 grid superimposed.
The name of a the gene is provided jointly with the matrix and used to select the row to be plotted.
}
\examples{
\dontrun{
plotGeneByName (gene="HOOK1", mets=falseLMet, expresmatrix=falseLExpr, fileName=NULL)
}

}
\keyword{gene}
\keyword{name}
\keyword{plot}
