% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreGenesMat.R
\name{scoreGenesMat}
\alias{scoreGenesMat}
\title{scoreGenesMat}
\usage{
scoreGenesMat(mets, expres, x1 = 1/3, x2 = 2/3, y1 = NULL,
  y2 = NULL, percY1 = 1/3, percY2 = 2/3, aReqPercentsMat,
  aWeightMifL, aWeightMifNonL)
}
\arguments{
\item{mets}{Matrix of methylation values}

\item{expres}{Matrix of expression values}

\item{x1, x2}{Coordinates of vertical points in the X axis. Because it is expected to contain methylation values that vary between 0 and 1 the default values are 1/3 and 2/3.}

\item{y1, y2}{Coordinates of vertical points in the Y axis. Leaving them as NULL assigns them the percentiles of yVec defined by `percY1` and `percY2`.}

\item{percY1, percY2}{Values used to act as default for `y1`and `y2` when these are set to `NULL`}

\item{aReqPercentsMat}{Matrix of minimum maximum percentage of counts to have in a given cell}

\item{aWeightMifL}{A matrix of weights to score the previous counts if the scatterplot has been classified as L.}

\item{aWeightMifNonL}{A matrix of weights to score the previous counts if the scatterplot has been classified as non-L}

\item{trace}{logical. Set to TRUE to print count matrices as they are computed}
}
\description{
\code{scoreGenesMat} scores scatterplots using a binary and a numeric schemes on a row-wise basis.
}
\examples{
\dontrun{
sampleSize <- dim(mets)[2]
numGenes <-   dim(mets)[1]
reqPercentages <- matrix (c(2, 20, 5, 1, 40, 20, 0, 1, 2), nrow=3, byrow=TRUE)
(theWeightMifL=matrix (c(2,-2,-sampleSize/5,1,0,-2,1,1,2), nrow=3, byrow=TRUE))
(theWeightMifNonL=matrix (c(0,-2,-sampleSize/5,0,0,-2,0,0,0), nrow=3, byrow=TRUE))
scoreGenesMat <- function(mets, expres,
                         x1=1/3, x2=2/3,
                         y1=NULL, y2=NULL, percY1=1/3, percY2=2/3,
                         aReqPercentsMat = reqPercentages,
                         aWeightMifL= theWeightMifL,
                         aWeightMifNonL= theWeightMifNonL) # , trace=FALSE)}
}
