% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toReqMat.R
\name{toReqMat}
\alias{toReqMat}
\title{toReqMat}
\usage{
toReqMat(numPoints, aReqPercentMat)
}
\arguments{
\item{numPoints}{Number of points in a scatterplot. Used to turn the required
percentages into required counts.}

\item{aReqPercentMat}{Matrix of required percentages.}
}
\value{
a counts matrix
}
\description{
\code{toReqMat} can be used to turn a matrix of required percentages into
a matrix of required counts to facilitate its scoring.
}
\examples{
reqPercentages <- matrix(
    c(
        3, 20, 5,
        5, 40, 20,
        4, 1, 2
    ),
    nrow = 3, byrow = TRUE
)
numberOfPoints <- 100
reqMat <- toReqMat(
    numPoints = numberOfPoints,
    aReqPercentMat = reqPercentages
)

}
